+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ## [Unreleased]
 
 ### Chore
+- formatting
 - comment codes for documentation
 
 ### Feat